requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
boto3==1.34.160
    # via -r requirements.in
botocore==1.34.160
    # via
    #   boto3
    #   s3transfer
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   typer
    #   uvicorn
distro==1.9.0
    # via openai
dnspython==2.6.1
    # via email-validator
email-validator==2.2.0
    # via fastapi
fastapi==0.112.0
    # via -r requirements.in
fastapi-cli==0.0.5
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   fastapi
    #   openai
idna==3.7
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.4
    # via fastapi
jiter==0.5.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mistune==3.0.2
    # via -r requirements.in
openai==1.40.3
    # via -r requirements.in
pillow==10.4.0
    # via -r requirements.in
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
pydantic-core==2.20.1
    # via pydantic
pygments==2.18.0
    # via
    #   -r requirements.in
    #   rich
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   uvicorn
rich==13.7.1
    # via typer
s3transfer==0.10.2
    # via boto3
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
starlette==0.37.2
    # via fastapi
tqdm==4.66.5
    # via openai
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   typer
urllib3==2.2.2
    # via botocore
uvicorn==0.30.5
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-cli
uvloop==0.19.0
    # via uvicorn
watchfiles==0.23.0
    # via uvicorn
websockets==12.0
    # via uvicorn
